Fix initial location puck missing
