+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/binder/map/RoutePreviewMapBinder.kt
@@ -7,7 +7,7 @@ import com.mapbox.navigation.dropin.DropInNavigationViewContext
 import com.mapbox.navigation.dropin.binder.Binder
 import com.mapbox.navigation.dropin.binder.navigationListOf
 import com.mapbox.navigation.dropin.component.camera.CameraComponent
-import com.mapbox.navigation.dropin.component.location.LocationPuck
+import com.mapbox.navigation.dropin.component.location.LocationComponent
 import com.mapbox.navigation.dropin.component.marker.LongPressMapComponent
 import com.mapbox.navigation.dropin.component.marker.MapMarkersComponent
 import com.mapbox.navigation.dropin.component.routeline.RouteLineComponent
@@ -19,7 +19,10 @@ internal class RoutePreviewMapBinder(
 
     override fun bind(value: MapView): MapboxNavigationObserver {
         return navigationListOf(
-            LocationPuck(value),
+            LocationComponent(
+                value,
+                navigationViewContext.viewModel.locationViewModel,
+            ),
             RouteLineComponent(value, navigationViewContext.routeLineOptions),
             CameraComponent(
                 value,
