+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/component/location/LocationViewModel.kt
@@ -7,11 +7,13 @@ import com.mapbox.android.core.location.LocationEngineResult
 import com.mapbox.geojson.Point
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
-import com.mapbox.navigation.core.trip.session.LocationMatcherResult
-import com.mapbox.navigation.core.trip.session.LocationObserver
+import com.mapbox.navigation.dropin.extensions.flowLocationMatcherResult
 import com.mapbox.navigation.dropin.lifecycle.UIViewModel
 import com.mapbox.navigation.ui.maps.location.NavigationLocationProvider
 import com.mapbox.navigation.utils.internal.logE
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
 
 sealed class LocationAction {
     data class Update(val location: Location) : LocationAction()
@@ -22,24 +24,24 @@ sealed class LocationAction {
 class LocationViewModel : UIViewModel<Location?, LocationAction>(null) {
     val navigationLocationProvider = NavigationLocationProvider()
 
+    /**
+     * Accessor function to get the last [Point]
+     */
     val lastPoint: Point?
         get() = navigationLocationProvider.lastLocation?.run {
             Point.fromLngLat(longitude, latitude)
         }
 
-    private val locationObserver: LocationObserver = object : LocationObserver {
-        override fun onNewRawLocation(rawLocation: Location) {
-            // We are not sending LocationAction.UpdateRaw here, because we are expecting to
-            // receive locationMatcherResult
+    /**
+     * Suspend until a non-null location is available.
+     */
+    suspend fun firstLocation(): Location {
+        var nonNullLocation: Location? = navigationLocationProvider.lastLocation
+        while (nonNullLocation == null) {
+            delay(DELAY_FIRST_LOCATION_MS)
+            nonNullLocation = navigationLocationProvider.lastLocation
         }
-
-        override fun onNewLocationMatcherResult(locationMatcherResult: LocationMatcherResult) {
-            navigationLocationProvider.changePosition(
-                location = locationMatcherResult.enhancedLocation,
-                keyPoints = locationMatcherResult.keyPoints,
-            )
-            invoke(LocationAction.Update(locationMatcherResult.enhancedLocation))
-        }
+        return nonNullLocation
     }
 
     override fun process(
@@ -70,11 +72,18 @@ class LocationViewModel : UIViewModel<Location?, LocationAction>(null) {
             }
         })
 
-        mapboxNavigation.registerLocationObserver(locationObserver)
+        mainJobControl.scope.launch {
+            mapboxNavigation.flowLocationMatcherResult().collect { locationMatcherResult ->
+                navigationLocationProvider.changePosition(
+                    location = locationMatcherResult.enhancedLocation,
+                    keyPoints = locationMatcherResult.keyPoints,
+                )
+                invoke(LocationAction.Update(locationMatcherResult.enhancedLocation))
+            }
+        }
     }
 
-    override fun onDetached(mapboxNavigation: MapboxNavigation) {
-        mapboxNavigation.unregisterLocationObserver(locationObserver)
-        super.onDetached(mapboxNavigation)
+    private companion object {
+        const val DELAY_FIRST_LOCATION_MS = 100L
     }
 }
